refactor: components/ResponsiveFooter/ に移動
